test(headless): address A/B test boundary review
