+++ packages/headless/src/__tests__/helpers/hash-fixture.ts
@@ -1,0 +1,5 @@
+import { createHash } from 'node:crypto';
+
+export function sha256(value: string): string {
+  return `sha256:${createHash('sha256').update(value).digest('hex')}`;
+}
